perf: update load order of CDN elements
This commit updates the load order of the CDN custom elements to ensure
that primitive elements are loaded BEFORE any element that has
dependencies on primitive elements loads their listed version.

The issue to solve is that elements that have listed dependencies
typically have older versions listed. It's this older version that is
then loaded into the client and disallowing the newer primitive from
ever being loaded into the client.

Changes to be committed:
modified:   public/index.html