docs: update FAQ content
These updates updates auro-hyperlink target attr refs to remove an issue
on the home and Support pages where the UI was unresponsive on render.

This issue may be a bug with auro-hyperlink where `relative` and
`target` are using on the same anchor tag.

Changes to be committed:
modified:   src/content/markdown/auroSupport.md
modified:   src/content/markdown/home.md